--- packages/nextra/src/page-map-cache.ts
@@ -1,26 +1,0 @@
-import type { FileMap, PageMapItem } from './types'
-
-class PageMapCache {
-  cache: {
-    items: PageMapItem[]
-    fileMap: FileMap
-  } | null = {
-    items: [],
-    fileMap: Object.create(null)
-  }
-
-  set(data: { items: PageMapItem[]; fileMap: FileMap }) {
-    this.cache!.items = data.items
-    this.cache!.fileMap = data.fileMap
-  }
-
-  clear() {
-    this.cache = null
-  }
-
-  get() {
-    return this.cache
-  }
-}
-
-export const pageMapCache = new PageMapCache()
