+++ packages/nextra/src/page-map.ts
@@ -90,3 +90,28 @@ export function resolvePageMap({
     dynamicMetaItems
   }
 }
+
+class PageMapCache {
+  cache: {
+    items: PageMapItem[]
+    fileMap: FileMap
+  } | null = {
+    items: [],
+    fileMap: Object.create(null)
+  }
+
+  set(data: { items: PageMapItem[]; fileMap: FileMap }) {
+    this.cache!.items = data.items
+    this.cache!.fileMap = data.fileMap
+  }
+
+  clear() {
+    this.cache = null
+  }
+
+  get() {
+    return this.cache
+  }
+}
+
+export const pageMapCache = new PageMapCache()
